requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
requests==2.31.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.1
python-multipart==0.0.6
jinja2==3.1.2
aioredis==2.0.1
asyncpg==0.29.0
Pillow==10.1.0
